target/arm: move arm_cpu_do_interrupt from tcg to common code
of note, during this move we replace a call to the helper:

helper_rebuild_hflags_a64(env, new_el);

with a call to arm_rebuild_hflags(); since helpers are not
available for KVM.

This _should_ be fine because pstate has been written already,
so current el etc should be up-to-date.

XXX to review with care.

Signed-off-by: Claudio Fontana <cfontana@suse.de>